paligemma/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
accelerate>=0.24.0
datasets>=2.14.0
Pillow>=9.5.0
numpy>=1.24.0
pandas>=2.0.0
tokenizers>=0.15.0
sentencepiece>=0.1.99
opencv-python>=4.8.0
albumentations>=1.3.0
PyYAML>=6.0
omegaconf>=2.3.0
hydra-core>=1.3.0
python-dotenv>=1.0.0

wandb>=0.16.0
tensorboard>=2.15.0
tqdm>=4.65.0

evaluate>=0.4.0
rouge-score>=0.1.2
bert-score>=0.3.13
sacrebleu>=2.3.1

scipy>=1.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
ipywidgets>=8.0.0
jupyter>=1.0.0

bitsandbytes>=0.41.0  
peft>=0.6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
pytest>=7.4.0

memory-profiler>=0.61.0
psutil>=5.9.0
google-colab
ipython